Api/auth

+++ social-media/Repository/UserRepository/UserRepository.cs
@@ -9,42 +9,53 @@ namespace social_media.Repository.UserRepository
 {
     public class UserRepository(DataContext context, IMapper mapper) : IUserRepository
     {
-
-        public async Task<List<User>> GetAll()
+        public async Task<List<UserDto>> GetAll()
         {
             var users = await context.Users.ToListAsync();
-            return users;
+            return mapper.Map<List<UserDto>>(users);
         }
 
-        public async Task<User> Get(Guid id)
+        public async Task<UserDto> Get(string id)
         {
             var user = await context.Users.FindAsync(id);
-            return user;
+            if (user == null)
+            {
+                throw new Exception("User not found");
+            }
+            return mapper.Map<UserDto>(user);
         }
-        public async Task<User> Create(User user)
+        public async Task<UserDto> Create(User user)
         {
             await context.Users.AddAsync(user);
             await context.SaveChangesAsync();
-            return user;
+            return mapper.Map<UserDto>(user);
         }
-        public async Task<User> Update(Guid id, [FromBody] UpdateUserDto updateUserDto)
+        public async Task<UserDto> Update(string id, [FromBody] UpdateUserDto updateUserDto)
         {
-            var user = await Get(id);
-            user.Username = updateUserDto.Username;
-            user.Password = updateUserDto.Password;
-            user.IsPublic = updateUserDto.IsPublic;
-            user.IsContentCreator = updateUserDto.IsContentCreator;
-
+            var user = await context.Users.FindAsync(id);
+            if (user == null)
+            {
+                throw new Exception("User not found");
+            }
+            var mappedUser = mapper.Map<User>(user);
+            mappedUser.UserName = updateUserDto.UserName;
+            mappedUser.PasswordHash = updateUserDto.PasswordHash;
+            mappedUser.IsPublic = updateUserDto.IsPublic;
+            mappedUser.IsContentCreator = updateUserDto.IsContentCreator;
             await context.SaveChangesAsync();
            
-            return user;
+            return mapper.Map<UserDto>(mappedUser);
         }
-        public async Task<User> Delete(Guid id)
+        public async Task<UserDto> Delete(string id)
         {
-            var user = await Get(id);
+            var user = await context.Users.FindAsync(id);
+            if (user == null)
+            {
+                throw new Exception("User not found");
+            }
             context.Users.Remove(user);
             await context.SaveChangesAsync();
-            return user;
+            return mapper.Map<UserDto>(user);
         }
 
     }
